pyproject.toml
[project]
name = "my_workflow_three"
version = "0.1.0"
description = "my_workflow_three using AI Foundry and Semantic Kernel"
authors = [
    { name = "Your name", email = "email@example.com" },
]
requires-python = ">=3.10,<=4.0.0"
dependencies = [
    "semantic-kernel[azure]>=1.29.0,<2.0.0",
    "azure-search-documents>=11.6.0b4",
    "azure-identity (>=1.22.0,<2.0.0)",
    "aiohttp (>=3.11.18,<4.0.0)",
    "pydantic>=2.11.4"
]

[project.scripts]
start = "my_workflow_three.main:run"
serve = "my_workflow_three.main:run_dapr"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.aifroundry]
type = "workflow"